+++ app/src/renderer/components/common/TmBtnCopy.vue
@@ -10,7 +10,20 @@ import { clipboard } from "electron"
 import { TmBtn } from "@tendermint/ui"
 export default {
   components: { TmBtn },
-  props: [`value`, `title`, `body`],
+  props: {
+    title: {
+      type: String,
+      required: true
+    },
+    body: {
+      type: String,
+      required: true
+    },
+    value: {
+      type: String,
+      required: true
+    }
+  },
   computed: {
     notifyTitle() {
       if (this.title) return this.title