+++ app/src/renderer/components/common/TmModalError.vue
@@ -1,9 +1,9 @@
 <template lang="pug">
 .tm-modal-error__wrapper
   .tm-modal-error
-    .tm-modal-error__icon: i.material-icons {{ errorIcon }}
-    .tm-modal-error__title {{ errorTitle }}
-    .tm-modal-error__body {{ errorBody }}
+    .tm-modal-error__icon: i.material-icons {{ this.icon }}
+    .tm-modal-error__title {{ this.title }}
+    .tm-modal-error__body {{ this.body }}
     .tm-modal-error__footer
       tm-btn#tm-modal-error__btn-issue(
         size="lg"
@@ -26,26 +26,26 @@ import { TmBtn } from "@tendermint/ui"
 export default {
   name: `tm-modal-error`,
   components: { TmBtn },
-  props: [`icon`, `title`, `body`],
+  props: {
+    title: {
+      default: `Voyager ran into an error`,
+      type: String
+    },
+    body: {
+      default: `Voyager has encountered a critical error that blocks the app from running. Please create an issue and include a copy of the app logs.`,
+      type: String
+    },
+    icon: {
+      default: `error_outline`,
+      type: String
+    }
+  },
   data: () => ({
     logPath: ``,
     issueUrl: `https://github.com/cosmos/voyager/issues`
   }),
   computed: {
-    ...mapGetters([`config`, `lastHeader`]),
-    errorIcon() {
-      if (this.icon) return this.icon
-      else return `error_outline`
-    },
-    errorTitle() {
-      if (this.title) return this.title
-      else return `Voyager ran into an error`
-    },
-    errorBody() {
-      if (this.body) return this.body
-      else
-        return `Voyager has encountered a critical error that blocks the app from running. Please create an issue and include a copy of the app logs.`
-    }
+    ...mapGetters([`config`, `lastHeader`])
   },
   mounted() {
     this.logPath = remote.getGlobal(`root`) + `/main.log`