+++ app/src/renderer/components/common/TmModalSearch.vue
@@ -38,7 +38,12 @@ export default {
     TmFormGroup,
     TmFormMsg
   },
-  props: [`type`],
+  props: {
+    type: {
+      type: String,
+      required: true
+    }
+  },
   computed: {
     ...mapGetters([`filters`, `lastHeader`]),
     open() {